requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
black==24.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.5.0
frozendict==2.4.0
idna==3.6
mypy-extensions==1.0.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
platformdirs==4.2.0
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
pytz==2023.4
requests==2.31.0
ruff==0.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
